requirements.txt
fonttools
pyoxipng
